[politeiad] Don't be clever at startup with anchors
Instead of trying to be clever and drop an anchor on restart of the
daemon just let the anchor code run as scheduled.  This prevents a race
when the daemon is (re)started right around the scheduled anchor drop
time.  What this race resulted in was a double logging of the same
anchor because the anchor was not confirmed yet and therefore the
database did not know it had dropped an anchor prior.