Add Decred plugin that enables voting on proposals.
Add infrastructure to do passthrough commands www to politeiad plugins.
The idea here is to ensure that there is no signing capable authority
directly connected to the internet.

This commit also adds a new tool to perform voting on proposals called
politeiavoter.

There are a number of XXX in the code that still require attention but
we are moving this to master to commence live testing.

This work was done with help from @sndurkin

Note that there are 2 tests that were commented out that need to be
added back.