[www] Fix bugs with GET handlers. (#219)
The GET handlers are trying to unmarshal the request body
as JSON into a struct, but a payload within a GET request
is not actually defined by the latest HTTP/1.1 spec. This
commit changes the behavior to instead pull values from
the query and store them in the command struct.