[www] Add user paywall verification route.
This new route verifies whether the given transaction is found on the
blockchain, is sent to the correct address, has the minimum number of
confirmations (2), comes after a fixed timestamp, and is of at least
the minimum amount of DCR (0.1). If all those are true, it sets the
transaction on the user in the database, marking the user as paid.