gitbe: use batch transactions dcrdata request
Adds batchTransactions, which uses dcrdata's new /txs (or
/txs/trimmed) endpoints.  It creates POST request body using the
dcrdata.Txns type.  The response is unmarshalled into a
[]dcrdataapi.TrimmedTx.
Adds largestCommitmentAddresses, which uses batchTransactions to
fetch all transaction info for a batch of ticket hashes.
Adds validateVoteByAddress, which takes the best commitment address
for the specified ticket.  This is obtained by largestCommitmentAddress
or largestCommitmentAddresses.
Modify pluginBallot to use largestCommitmentAddresses and
validateVoteByAddress.

dep ensure update dcrdata