[d+www] Add vote parameters to start vote. (#552)
This diff adds the parameters QuorumPercentage and PassPercentage to
the start vote command and makes the vote duration a configurable www
parameter. QuorumPercentage is used to indicate the percentage of
eligible tickets that must vote in order for a quorum to have been
met. PassPercentage is used to indicate the percentage of total votes
that are required for the vote to pass.