www: Fix NumOfProposals value in UserProposals
This diff fixes how the NumOfProposals value is calculated in the
UserProposals endpoint. Before, the NumOfProposals value was the total
number of proposals that had been submitted by the user regardless of
proposals status. Now, the NumOfProposals value reflects the number of
proposals that are availble to be returned from the UserProposals
endpoint. If the request is being sent by an admin or a user is
requesting their own proposals then NumOfProposals includes unvetted,
censored, and public proposals. Otherwise, NumOfProposals only includes
public proposals.