d+www: Add the proposal status abandoned (#599)
* www: Catagorize proposals by state

* d+www: Add the proposal status abandoned

This commit adds the 'archived' record status to politeiad and the
'abandoned' proposal status to politeiawww. An abanonded proposal in
politeiawww corresponds to a archvied proposal in politiad. Admins can
declare a proposal abandoned at any time as long as the proposal is
public and voting has not yet started. The admin must include a change
message when changing a proposal status to abandoned and the admin's
action is logged. Abandoned proposals are publicly viewable but cannot
be edited, have the voting period authorized for them, have any new
comments made or have any new comment votes cast.  Abandoned proposals
are vetted proposals and are returned from the politeiad and politeiawww
endpoints for fetching vetted proposals.