multi: Add cache layer (#660)
* multi: Add cache layer

This commit adds a cache layer to politeiad and politeiawww.  The
politeiad git repositories serve as the source of truth for proposal
data and a CockroachDB database acts as a cache for proposal data.

politeiad has read and write access to the cache.  politeiawww has only
read access to the cache.  The flow of data is as follows:

1. politeiawww receives a command from a user
2. politeiawww creates a politeiad request for the command and sends it
3. politeiad writes the command data to the git repository and updates
   the cache
4. politeiad returns a response to politeiawww
5. politeiawww reads the cache if needed
6. politeiawww returns a response to the user

The cache is automatically rebuilt on cache version change.  You can
also force the cache to rebuilt by using the flag --buildcache on
politeiad startup.

The politeiad record and all record metadata has been removed from the
politeiawww inventory.  The only thing left in the inventory is decred
plugin data.  The decred plugin data will be added to the cache and
removed from the inventory in a future PR.

Since this commit touched so much of the code for the `/proposals`
endpoints, I began cleaning up `backend.go` and moving the updated
`/proposals` enpoint code to `proposals.go`.

I had to delete all of the tests that relied on the proposal record in
the inventory.  I re-wrote most of these tests in `proposals_test.go`.

* Add vote data to cache

* build plugin cache and fix stuff

* cleanup

* doco and cleanup

* vim playing tricks on me

* readme update

* doco fixes

* votestatus bug fix

* fix nits

* doco fix