politeiad: Fix new comment error bug (#684)
This commit fixes a bug that was causing an error to be incorrectly
logged when a comment is submitted to a proposal that already has
existing comments.

The error is only supposed to be logged if the politeiad in-memory
comments cache contains an existing comment with the same commentID as
the new comment.  The error is actually being logged anytime there are
any existing comments in the cache for the given record.