wwwcli: Add testrun cmd and other improvements
This commit makes various improvements to politeiawwwcli:

* Adds some missing commands: activevotes, votestatuses, and
userproposalcredits.

* Updates how user identities are stored so that you no longer need to
run `updateuserkey` when switching between users.  All users identities
are now stored and the logged in user's identity is automatically
loaded.

* Performs a large style cleanup.  All golangci-lint errors have been
fixed for politeiawwwcli that were not false positives.

* Adds the command testrun.  This command performs a series of high
level tests on the politeiawww routes.

In addition to politeiawwwcli improvements, this commit also:

* Gets politeiawww_dataload working again.

* Deletes the politeiawww_refclient.sh script.