Add NewInvoice Route and cmsdb to store data
This is the first commit to receive and store invoices as records.

Invoices are treated very similiarly to proposals, just with different
metadata.  They are both sent to politeiad and stored as records.

Now since we need to store various information about invoices beyond
record information, we have added another cockroachdb that is contained
within the database package.  It currently just contains Invoice and
LineItem tables.  In the future we will be adding more tables as needed.
These Invoice and LineItem tables should be able to be built and recovered
from an record Inventory call to politeiad.

Next we will be adding various ways to obtain information from the cmsdb.
Administrators should be able to search by date ranges, userid and status.