www: Add proposal tests
This commit adds a few things related to politeiawww testing.

- Adds a testpoliteiad package.  testpoliteiad is a httptest server that
implements the politeiad api.

- Adds a testcache package.  testcache is an implementation of the Cache
interface that stores records in memory and that can be used for
testing.

testpoliteiad and testcache allow us to fake the responses from
politeiad and the cache so that we can fully test the politeiawww
routes.

This commit also adds test coverage to processNewProposal,
setProposalStatus, and verifyStatusChange.

This commit increases politeiawww test coverage from 15.7% to 21.6%.