www: Fix fireEvent bugs (#730)
The _fireEvent function was being called without the lock being held in
a few different places.  This commit updates those calls to use the
fireEvent function instead, which does not require the caller to have
the lock held.

It also fixes an error in util/paywall.go that travis was complaining
about.