d: Remove requirement on pre-module dcrdata
dcrdata currently depends on politeia, which depends on a pre-module release of
dcrdata.  Because this earlier version does not follow semantic import
versioning, it is impossible for dcrdata to currently create submodules starting
at v1.0.0.  Updating politeia to a module-aware dcrdata will fix this dcrdata
build issue after requiring the updated politeia.