Add InviteNewUser and Register routes for CMS (#710)
This adds the first of many upcoming routes for cmswww mode.

The Version request will return the current mode that politeiawww is
running (either piwww or cmswww).

Administrators will need to invite any new users and the user will be
sent an email to provide the user with a verification token.  The users
will then be able to register for the Contractor Management System.