Add UserInvoices and AdminInvoices routes (#756)
This commit adds the UserInvoices and AdminInvoices routes that allow
for users and administrators to request invoices based on certain
criteria.

Currently UserInvoices simply responds with all of the currently logged
in user's invoices. While the AdminInvoices request requires an admin
to be logged in.  AdminInvoices allows searching by month/year and/or
status.  If none of these fields are used, then all of the current
invoices will be returned.