Add EditInvoice and SetInvoiceStatus routes (#753)
EditInvoice allows users to update their invoices based on comments by
administrators.  This will change the status of the invoice back to
InvoiceStatusUpdated so the administrators will be notified that the
invoice must be reviewed again.

SetInvoiceStatus allows administrators to update the status of a given
invoice.  This can be set to updated, disputed, rejected or approved.

For either of these commands, a corresponding request is sent to the
politeiad backend to update the records and keep the updates journaled.