www: Add TokenInventory route
This diff adds a TokenInventory route to politeiawww.  The
TokenInventory route retrieves the censorship record tokens of all
proposals in the inventory.  The returned tokens are categorized in a
manner that resembles their vote statuses, but that has been altered to
be a more natural fit with how a UI would display them.  The categories
are pre, active, finished, and abandoned.

The 'finished' category was not broken up into 'approved' and 'rejected'
because politeia is designed to be able to handle proposal votes with
multiple voting options.  Categorizing the finished proposal votes into
'approved' and 'rejected' only makes sense for simple yes/no proposal
votes.

The cache StartVote model was also updated.  The type of the EndHeight
field was changed from a string to a uint64 so that, given the best
block, a query can categorize the results into 'active' and 'finished'
voting periods.

Query performance on my local machine when testing against the mainnet
pi production data was as follows:

```
pre query           : 98ms
active query        : 97ms
finished query      : 102ms
abandoned query     : 4ms
```

Keep in mind that my local machine performance is significantly slower
than the production pi performance.