[cms] Add new comment endpoint for invoices (#790)
This adds a new NewCommentInvoice endpoint for cms. Having a
fully separate version was done to avoid the paywall check and the
proposal specific checks there.

Currently, only admins or the owner/creator of a given invoice should
be able to comment on that invoice. And it must not be in Approved
or Paid status.