[cms] Add ExchangeRate to invoice json
This adds an ExchangeRate field to the InvoiceInput struct.

With this change it requires newly submitted invoices to have
the corresponding month's exchange rate included in the invoice
payload.  In doing so, it requires the cli request for newinvoice to
first request the exchange rate from the server to use.  A similar
process will be required when using the GUI.

This also adds additional fields to the generated payouts for admins.
Exchange rate and the total DCR to be paid out by the invoice.