dbutil: Add stub users command
This commit adds the --stubusers option to the politeiawww_dbutil tool.
The stub users option creates user stubs in the user database for the
public keys found in the passed in directory.  It is intended to be used
in conjunction with the politeiaimport tool to allow developers to
import the decred mainnet proposal data and test against it.