[cms] Add invoice exchange rate endpoint (#806)
This is a first pass of implementing an exchange rate endpoint.
It calculates a monthly average exchange based on a provided month/year.

Currently, this is calculation is done each time. Once this implementation
and calculation has been approved/merged, I will go about adding a caching system
in the cms database that will only request the prices and calculate when if there
is not an entry found in the database.