multi: Add token inventory approved and rejected
This commit removes the 'finished' field from the token inventory reply
and replaces it with the fields 'approved' and 'rejected'. These fields
correspond to proposals with approved and rejected votes.

A cache table for vote summaries was added to enable proposal lookups by
whether their vote was approved or rejected. The vote summaries table is
lazy loaded. Loading vote summaries required the addition of a new
decred plugin command since only politeaid has write access to the
cache.