d: Sets valid mime types through config file
This PR allows for valid mime types to be setted on config file.
It removes the obsolete v1/mime package, and moves its functions
to util/file. In order to accomplish this, the function that ve-
rifies censorship record files had to be moved also, which makes
d's api/v1 cleaner and with single responsibility, that of defining
types, and not functions.