gitbe: Fix duplicate votes bug
This commit fixes a bug in the decred plugin ballot command that would
allow duplicate votes to be recorded in the votes journal in specific
circumstances.

politeiad has a memory cache of all cast votes. When a user casts a
vote, politeiad checks the memory cache to see if the vote already
exists. If it does not exist, the vote is writen to the votes journal
then added to the memory cache. This memory cache read/write operation
was not atomic. If a user submitted the same vote request multiple
times, concurrently, it was possible for a duplicate vote to be added to
the votes journal before the memory cache was properly updated.