cache: Remove duplicate vote lookup
This commit updates how votes are inserted into the cache. It removes
the vote lookup that was checking for duplicate votes. This lookup query
was too expensive and is not needed. It was causing cache build times to
be unreasonable (~3 hours) and would also cause extremely long response
times if someone were to cast a large block of votes all at once, which
does occasionally happen.

Removing this query does not effect whether duplicate votes will be
added to cache once they have been rejected by gitbe. Any vote that does
not have a valid vote receipt is not added to cache. Duplicate votes
that are rejected by gitbe do not contain a valid vote receipt.

Removing this query does mean that the 15 duplicate votes that made it
into the "Decentralize Treasury Spending" proposal vote journal will be
added to the cache and reflected in the vote total. I don't consider
this to be a huge issue since the proposal passed with a 97.49% approval
rating. Documenting the 15 duplicate votes in the Politeia Digest and
Decred Journal should be sufficient.