[cms] Add DCC Proposal System
This adds the initial implementation of the DCC issuance and revocation
proposal system.

Stakeholder Approved DCC Proposal:
https://proposals.decred.org/proposals/fa38a3593d9a3f6cb2478a24c25114f5097c572f6dadf24c78bb521ed10992a4

All submitted DCC proposals will be processed in a very similar fashion
as proposals and invoices.  They will all be of an expected, versioned
form that will allow them to be submitted as records to politeiad.  This
allows us to use the politeiad backend without any required changes.
And another added benefit of that is we are able to reuse the existing
commenting system to allow for discussions on DCC proposals, as well as
users showing support or opposition of a given DCC proposal.

Authenticated Users are now provided with the following requests:
- NewDCC (/dcc/new)
  Creates a new DCC proposal for issuances or revocations.  Proposals
  must include a statement supporting the nominated user,
- NewDCCUser (/dcc/newuser)
  Creates a new "dummy" user that will be nominated in a coming DCC.
- SupportDCC (/dcc/support)
  Creates a comment on the DCC that simply says "aye."
- OpposeDCC (/dcc/support)
  Creates a comment on the DCC that simply says "nay."
- GetDCC (/dcc/token)
  Returns the DCC given the included token.
- GetDCCs (/dcc/status)
  Returns all DCCs that match the included status.
- NewCommentDCC (/dcc/comment)
  Creates a comment on a DCC (that isn't "aye" or "nay").

Additionally there are 2 new requests for administrators:
- ApproveDCC (/admin/approvedcc)
  Approve a DCC and thereby invite the nominated user to CMS (using
  existing InviteNewUser processing).
- RejectDCC (/admin/rejectdcc)
  Reject a DCC and thereby change the user's status that will cause
  their CMS access to be revoked or altered.

The initial plan is to have administrators give the final approval or
rejection for a DCC based on final Support/Opposition tally and
comments.  Then as other subsystems are put into place, we can begin to
develop the "all contractor vote" system.  This will used in instances
where a DCC is considered contentious or in dispute.