www: Remove email lookups from reset password
This diff removes the user lookups by email address from the reset
password route and splits the ResetPassword route into two separate
routes.

ResetPassword
VerifyResetPassword

This is consistent with the pattern that is used for VerifyNewUser and
VerifyUpdateUserKey.

Since the ResetPassword command now contains the username and email
address, it was necessary to add a minimum wait time to the route in
order to prevent an attacker from being able to execute a timing attack
to determine if the provided email address is the user's valid email
address.