tlog: Add reference client and server implementation.
This diff adds the tlog package which provides a reference tlog client and server implementation.  

tserver provides an API for storing arbitrary records and timestamping those records onto the Decred blockchain.  It stores the records on disk and uses Google's open source project trillian to store a hash of the record in an append only log where the merkle root of each log is episodically timestamped onto the Decred blockchain.

tclient provides a client for interacting with the tserver API.