multi: Prevent duplicate comments.
This diff prevents a user from submitting a duplicate comment. A
duplicate comment is defined as a comment with the same signature as an
existing comment.  The comment signature is the token+parentID+comment.

The scenerio that this prevents is when a user submits the same comment
request multiple times due to network and/or browser hiccups. This has
happened to multiple users on production pi.