multi: Prevent duplicate comments.
This diff adds validation to politeiawww to prevent a user from
submitting a duplicate comment. A duplicate comment is defined as a
comment with the same signature as an existing comment. The comment
signature is the token+parentID+comment.

The scenerio that this is meant to prevent is when a user submits the
same comment request multiple times due to some network/browser 
hiccups. This has happened to multiple users on production pi.

It also cleans up some documentation that was out of date in 
www/v1/api.md.

Note: this only adds validation to politeiawww. It does not add any
validation to politeiad. This means thats its still possible to for
duplicate comments to make it into the politeiad repo if the duplicate
comment request is submitted before the original comment is added to the
cache. I explored the possibility of adding duplicate comment validation
to politeiad, but doing so would make the politeiad new comment call
much more expensive and I don't think the tradeoff is worth it.