[cms] Add Support, Oppose and Comment DCC requests
This PR adds the ability for users to offer their support, opposition
and comment on active DCC proposals (issuances or revocations).

When a user offers there support or opposition of a given DCC proposal,
metadata containing user information is appended on the existing
journal entry on the politeiad backend.  This is similar to the process
of changing the status of a proposal.  All changes are tracked and
immutable once attached to a journal entry.