[cms] Add DCC Proposal System foundation (#980)
This adds the initial implementation of the DCC issuance and revocation
proposal system.

Stakeholder Approved DCC Proposal:
https://proposals.decred.org/proposals/fa38a3593d9a3f6cb2478a24c25114f5097c572f6dadf24c78bb521ed10992a4

All submitted DCC proposals will be processed in a very similar fashion
as proposals and invoices.  They will all be of an expected, versioned
form that will allow them to be submitted as records to politeiad.  This
allows us to use the politeiad backend without any required changes.
And another added benefit of that is we are able to reuse the existing
commenting system to allow for discussions on DCC proposals.

Authenticated Users are now provided with the following requests:
- NewDCC (/dcc/new)
  Creates a new DCC proposal for issuances or revocations.  Proposals
  must include a statement supporting the nominated user.
- DCCDetails (/dcc/{token})
  Returns the DCC given the included token.
- GetDCCs (/dcc)
  Returns all DCCs that match the included status.

There will be subsequent PRs that will provide further DCC functionality,
including: Supporting/Opposing DCCs, Commenting on DCCs and Admin
Approval and Rejection.