[cms] Add Support, Oppose and Comment DCC functionality (#981)
This PR adds the ability for users to offer their support, opposition
and comment on active DCC proposals (issuances or revocations).

When an authorized user submits a Support or Opposition request,
metadata from the request is appended to the existing DCC on
the politeiad backend. This makes them permanent and verifiable.
This PR adds the ability for users to offer their support, opposition
and comment on active DCC proposals (issuances or revocations).