politeiavoter: Improve eligible vote handling.
This modifies the way vote eligibility is determined to ignore
commitments for imported xpubs, refactors the relevant vote eligibility
determination code into a separate function, and updates the inventory
path to report eligible remaining votes in addition to the eligible
tickets.

Credits to @jrick for the code that ignores commitments for imported
xpubs.