[cms] Add SetDCCStatus request (#982)
This PR completes the initial implementation of the DCC system.

This allows administrators to manually approve and reject DCC proposals.
This process will be somewhat automated in the future, but until then
will require administrator's authorization due to the high degree of
possible error in the process.

Upon approval of an issuance, the user will be have their ContractorType
updated to Direct or SubContractor. This will now allow them (or a supervisor)
to submit invoices on their behalf. Upon approval of a revocation, the user will
have their ContractorType updated to Revoked. This will make it so they are
unable to submit subsequent invoices. We may enact further login restrictions
in the future, but that's still up for debate.

The next step in the process will be implementing the "all contractor
vote" in the event that a particular DCC has been found to be "debatable"
or not easily approved/rejected by an administrator