www: Split politeiawwwcli into piwww and cmswww.
The politeiawwwcli cmd was getting bloated and hard to use due to it
containing commands for both piwww and cmswww. This is particulary true
for commands that shared the same route but had different responses,
like the userdetails command. Since the responses are different it
required that you have a pi user details command (userdetails) and a
seperate cms user details command (cmsuserdetails). Having to know when
a cms specific command was required vs when you could use the same
command for both pi and cms was making politeiawwwcli difficult to use.

This diff splits the politeiawwwcli cmd into three seperate packages:

piwww - CLI tool for interacting with the piwww API.
cmswww - CLI tool for interacting with the cmswww API.
shared - library that contains the commands used by both piwww and
cmswww.

The piwww tool only contains the commands for the pi API and the cms
tool only contains the commands for the cms API. The shared package
ensures that there is no duplicate code.

Note: piwww and cmswww now use their own config files.  See the README
in each package for instructions on setting up the config file.