multi: Cleanup batched calls in cache and userdb.
This diff standardizes how batched called are handled in the cache and
the userdb. Before this diff, some batched calls where returning a slice
of results and some were returning a map of results. Having all batched
calls perform the same way makes the code easier to reason about.

All batched calls in the cache and the userdb now return a map of the
results. If an output is not found for one of the input values, such as
if a passed in proposal token doesn't actually correspond to a proposal,
the input value will not be included in the returned map. This makes it
easy for the calling function to check if results where found for all
input values and leaves it up to the calling function to decide how to
deal with values that were not found.