[pivoter] Make politeiavoter less prone to network errors (#1022)
The big idea is to make politeiavoter a fire and forget type deal even
when dealing with terrible connections over tor.

When a vote fails move it to a retry queue where the vote will be
retried randomly within 120s. Failed votes are pushed to the end to the
retry list. If a vote fails in the retry loop it is also pushed to the
back of the queue. Votes that error out on the server are not retried;
this only deals with network errors.

Add signal handler to dump votes that are in the retry queue. Sending a `SIGUSR1` to the `politeiavoter` process will result in a dump of all activity.

Autodetect vote trickle duration when `--voteduration` is set to zero. We now require the `--trickle` to be set in order to start the process.