Update dcrdata imports: - github.com/decred/dcrdata/api/types/v3 => github.com/decred/dcrdata/api/types/v4 - github.com/decred/dcrdata/pubsub/types/v2 => github.com/decred/dcrdata/pubsub/types/v3 - github.com/decred/dcrdata/pubsub/v2/psclient => github.com/decred/dcrdata/pubsub/v3/psclient
Unchanged types in api/types:
- BlockDataBasic
- Txns

api/types.TrimmedTx is the same except for the Vin field that returns a
slice of a dcrd/rpc/jsonrpc/types type instead of a dcrjson type.  But
that is OK because: (1) politeia (gitbe) does not use the Vin field,
and (2) the Vin field is identical.

In dcrdata's pubsub/psclient package, ping is no longer a subscription.
The server will ping the client automatically. Thus, in politeiawww,
the function (*wsDcrdata).subToPing is removed.

Unchanged types in pubsub/types used by politeia:
- AddressMessage
- TxList

Update dcrtime, which has no relevant code changes, but it does clean
up the dependencies.