multi: Add RecordStatusChangeV2 mdstream.
This diff adds a RecordStatusChangeV2 to the mdstream package and
updates politeiawww to use the RecordStatusChangeV2 mdstream.

RecordStatusChangeV2 adds a Signature field, which was erroneously left
out of V1. The status change signature was being included in the
politeiawww API call correctly and was being validated by politeiawww
correctly, but was not being saved to gitbe since it was missing from
the V1 mdstream. This diff fixes that.

No politeiawww API changes were required since the API was already
including the signature correctly.