[cms] Add tool to check for payment metadata (#1062)
Due to previously overlooking needing to add payment metadata
to the backend, invoices that were paid before #1051
are missing their backend metadata.

This tool simply looks up the paid payments from the db and checks
the invoice record for its metadata. If there is no metadata found, then
it submits it.