[cms] Rebuild cmsDB from records cache (#1065)
This PR now allows for rebuilding of the cmsdb based on db version
changes.

Adding the --buildcmsdb option to politeiawww will no drop relevant tables
in cmsDB (invoices, line items, payments, invoice changes and dccs). Then
it will take all of the records and metadata currently stored in the backend's cache
and rebuild the tables from that information.