multi: Move www api dependent code out of util.
This diff adds a VerifySignature() method to the StartVote struct so
that the signature and signature verification are explicitly documented
in code.

In order to add the VerifySignature() method, functions that referenced
the www v1 API had to be moved out of the util package due to a circular
dependency. These functions are all www specific so it makes sense for
them to not be in the util package anyway.

This diff is part of the work to fix the StartVote mdstream signature,
but I broke it out into a separate PR since it moves around a bunch of
unrelated code.