multi: Move www api dependent code out of util.
This diff adds a VerifySignature() method to the decredplugin StartVote
struct so that the signature is explicitly documented in the code.

In order to add the VerifySignature() method, functions that referenced
the www v1 API had to be moved out of the util package due to circular
dependency errors. All of the code that was moved was specific to
politeiawww and did not need to be in the global util package. The
websocket functions were moved into a www specific util package so that
they can be used by both politeiawww and the CLI tools.

This diff is part of the work to fix the StartVote mdstream signature,
but I broke it out into a separate PR since it moves around a bunch of
unrelated code.