politeiawww: Optimize ActiveVote route.
This diff optimized the ActiveVote route with newer decred plugin calls
that weren't around when the ActiveVote route was initially implemented.
It was previously fetching the entire politeiad inventory and the the
vote details for every proposal in the inventory. It now uses the
TokenInventory command to determine which proposals are actively being
voted on and only fetches the data for those specific proposals.