[cms] Add CMS Users Request (#1100)
This PR adds a new authenticated route to all users to request
a list of users in various domains and/or contractor types. If 
no domain or contractor type is given then the full list of 
contractors is provided. I don't believe any of the returned 
information should be considered private.

This is required for things like dropdowns and other UX 
improvements for DCC submissions.