www: Update bestblock cache using wsDcrdata.
This diff subscribes politeiawww to the newblock event when it is in www
mode and caches the current best block. This cached value is used in
requests where the current best block is needed, but the old method of
requesting it from politeiad each time is used as a fallback when there
is no active websocket connection.

This diff also adds a reconnect method to wsDcrdata that creates a new
websocket client, and resubscribes to all of the events the old client
was subscribed to. This new method is used in the case of a hang up
message or a read timeout from dcrdata.