politeiawww: Move sessions to userdb.
This diff moves user sessions from a session store backed by the
filesystem to a session store backed by the userdb. It implements a
custom sessions store that satisfies the gorrilla/sessions Store
interface.

Moving the session store to the userdb accomplishes a few things:

1. Allows us to manually invalidate sessions. An example being when a
   user changes their password we can invalidate their sessions on all
   other devices. This was not previously possible with the filesystem
   store implementation.

2. Removes one of the blocking issues that is preventing us from running
   multiple politeiawww instances.

This diff does not fix issues #647 or #650. It is a prerequisite for
fixing those issues, but those issues will be fixed in the next PR.