cache: Add ProposalGeneralMetadata table.
This diff adds the ProposalGeneralMetadata table to the cache in order
to make the ProposalGeneral mdstream queryable. This mdstream was
previously being stored in the cache as a MetadataStream with an encoded
payload.  It still is stored as a MetadataStream, but a
ProposalGeneralMetadata record will now exist for the most recent
proposal version. This will be required once RFP proposals go live so
that the link between proposals can be queried.

The diff also required adding hooks for cache plugins so the the
ProposalGeneralMetadata table could be updated properly when a proposal
is submitted or edited.