dbutil: Add verifyidentities command.
This diff adds a command to politeiawww_dbutil that verifies that a
user's identities do not violate any politeia rules. Any invalid
identities that are found are deactivated.

The bug that lead to these invalid states has already been fixed and the
code has been updated to automatically deactivate invalid identities.
This command allows us to do it manually until the code changes have
been deployed to production.