multi: Add v2 StartVote.
The issue that prompted this PR was that the StartVote signature was
only signing the proposal token when it should have been signing a hash
of the Vote struct, thus commiting to the provided vote options and vote
params. 

This diff makes the following changes to the decred plugin StartVote:
* A ProposalVersion field, which specifies the version of the proposal
  that is being voted on, was added to the Vote struct so that the
  proposal version is explicitly included in the StartVote signature.  
* A Type field was added to the Vote struct in order to specify the vote
  type. This will make it easier to add additional vote types, such as
  runoff votes for an RFP proposal, in the future.
* The StartVote Signature is now the signature of a hash of the Vote
  struct. It was previously the signature of just the proposal token.

These changes are not backwards compatible and required versioning both
the decred plugin StartVote and the the politeiawww API.

The following routes have been added to the politeiawww v2 API:
* StartVote: start a vote using the v2 StartVote.
* VoteDetails: fetch the vote details for the given proposal. This
  returns all of the information contained in a StartVote and
  StartVoteReply. The StartVote version is included in the response and
  the Vote in included as a JSON encoded string that must be decoded
  according to the version.

All v1 routes that return vote data (BatchVoteSummary, VoteResults,
ActiveVotes, VoteStatus) have remained unchanged and still function
normally.