[cms] Add Temporary contractor type (#1132)
This is the beginnings of the ability to add "one-off" contractors 
that are allowed to submit invoices without having to proceed 
through the full DCC process as a normal regularly billing 
contractor would.

The current use-case that comes to mind for these users are bug 
bounty folks. While they need the ability to bill for work complete, 
they are not considered to be full members of our contractor 
community and with that will come a reduce amount of rights 
and access.