[cms] Add codestats request
This PR allows administrators to now query codestats for any particular
user over a requested period of time.  This is done by connecting to
a populated cockroachdb that contains relational information about
pull requests, commits and reviews.

A request to update the github-tracker information will be added
in a subsequent PR. It will be completed during the same cronjob
that handles the fetching of exchange rate data at the beginning
of each month.