politeiawww: Handle cookie key change gracefully.
This diff updates the SessionStore to be able to handle a change to the
cookie key gracefully. Changing out the key used to encode session
values in the cookie would previously cause a 500 if anyone passed up a
session that was encoded by the previous key. If the existing session
value is not able to be decoded properly, the session is now considered
invalid and a new session must be created.